Initial commit of tool to generate usernames
It's fairly basic logic right now (and not written very efficiently),
but it can generate random, mostly human readable usernames.

The general idea is to combine consonants and vowels in a natural-like
way (albeit, I'm sure there's some tuning required).

It's not exactly going to generate a hardended high entropy username,
but that's okay since the goal of a random username is around defense
through obscurity + it generally not being possible to enumerate all
usernames on a service.